Move is_categorical_dtype to new location.
